# api/public/docs/openapi.yml
openapi: 3.1.0
info:
  title: 'Constellation APIs'
  license:
    name: 'GNU AFFERO GENERAL PUBLIC LICENSE'
    url: 'https://www.gnu.org/licenses/agpl-3.0.en.html'
  version: '2.0'
servers:
  -
    url: 'http://localhost:8080/api'
    description: 'The local environment.'
paths:
...
  /v1/service_versions:
    get:
      tags:
        - 'Service version'
      summary: 'Get a listing of the service versions'
      description: 'Get all service versions.'
      operationId: getServiceVersions
      parameters:
        -
          $ref: '#/components/parameters/base-filter-per-page'
        -
          $ref: '#/components/parameters/base-filter-page'
        -
          $ref: '#/components/parameters/base-filter-sort'
        -
          $ref: '#/components/parameters/base-filter-q'
        -
          name: 'filter[id]'
          in: query
          description: 'Filter by service version id.'
          schema:
            type: integer
        -
          name: 'filter[service_id]'
          in: query
          description: 'Filter by service id.'
          schema:
            type: integer
      responses:
        '200':
          description: 'Service versions list'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  data: { type: array, items: { $ref: '#/components/schemas/resource-service-version' } }
                  message: { type: string }
                  total: { type: integer }
                type: object
    post:
      tags:
        - 'Service version'
      summary: 'Store a service version'
      description: 'Store a service version.'
      operationId: storeServiceVersion
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/request-create-service-version'
      responses:
        '200':
          description: 'Created service version data.'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/resource-service-version' }
                  message: { type: string }
                type: object
        '422':
          description: 'Invalid inputs.'
          content:
            application/json:
              schema:
                properties:
                  errors: { properties: { '{name-of-invalid-field}': { type: array, items: { description: 'Error message', type: string } } }, type: object }
                  message: { type: string }
                type: object
  '/v1/service_versions/{id}':
    get:
      tags:
        - 'Service version'
      summary: 'Display the specified service version'
      description: 'Get a service version.'
      operationId: showServiceVersion
      parameters:
        -
          name: id
          in: path
          description: 'id of the service version'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Service version detail'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  data: { type: array, items: { $ref: '#/components/schemas/resource-service-version' } }
                  message: { type: string }
                type: object
        '404':
          description: 'Resource not found'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string }
                  exception: { type: string }
                  file: { type: string }
                  line: { type: integer }
                  trace: { type: array, items: { properties: { file: { type: string }, line: { type: integer }, function: { type: string }, class: { type: string }, type: { type: string } }, type: object } }
                type: object
    put:
      tags:
        - 'Service version'
      summary: 'Update an service version'
      description: 'Update an service version.'
      operationId: updateServiceVersion
      parameters:
        -
          name: id
          in: path
          description: 'id of the service version'
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/request-create-service-version'
      responses:
        '200':
          description: 'Updated service version data.'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/resource-service-version' }
                  message: { type: string }
                type: object
        '422':
          description: 'Invalid inputs.'
          content:
            application/json:
              schema:
                properties:
                  errors: { properties: { '{name-of-invalid-field}': { type: array, items: { description: 'Error message', type: string } } }, type: object }
                  message: { type: string }
                type: object
        '404':
          description: 'Resource not found'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string }
                  exception: { type: string }
                  file: { type: string }
                  line: { type: integer }
                  trace: { type: array, items: { properties: { file: { type: string }, line: { type: integer }, function: { type: string }, class: { type: string }, type: { type: string } }, type: object } }
                type: object
    delete:
      tags:
        - 'Service version'
      summary: 'Delete an service version'
      description: 'Remove the specified service version from storage.'
      operationId: deleteServiceVersion
      parameters:
        -
          name: id
          in: path
          description: 'id of the service version'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Service version deleted.'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  message: { type: string }
                type: object
        '404':
          description: 'Service version not found.'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  message: { type: string }
                type: object
...
components:
  schemas:
...
    request-create-service-version:
      title: 'Create service version'
      description: 'Create service version request'
      required:
        - version
        - service_id
      properties:
        version:
          description: Version
          type: string
        service_id:
          description: 'Service id'
          type: integer
      type: object
...
    resource-service-version:
      title: ServiceVersion
      description: 'Service version resource'
      properties:
        id:
          description: Id
          type: integer
        service_name:
          description: 'Service name'
          type: string
        service_id:
          description: 'Service id'
          type: integer
        version:
          description: Version
          type: string
        created_at:
          description: 'Creation date'
          type: string
          format: date-time
        updated_at:
          description: 'Last update date'
          type: string
          format: date-time
        meta:
          description: meta
          type: array
          items:
            type: string
      type: object
...
  parameters:
    base-filter-per-page:
      name: perPage
      in: query
      description: 'Number of items displayed by page.'
      schema:
        type: integer
        default: 100
        enum:
          - 5
          - 10
          - 25
          - 100
          - 1000
    base-filter-page:
      name: page
      in: query
      description: 'Current page.'
      schema:
        type: integer
    base-filter-sort:
      name: sort
      in: query
      description: 'Field used to sort the result.'
      schema:
        type: string
      examples:
        string:
          summary: '-id'
          value: '-id'
    base-filter-q:
      name: q
      in: query
      description: 'Query string'
      required: false
      schema:
        type: string
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Basic Auth'
      scheme: bearer
security:
  -
    bearerAuth: []